refactor: single mapping/detection sites for key groups, unique violations, video polling
- VirtualKeyGroupDto.FromEntity replaces six hand-written entity->DTO copies.
  Only the update endpoint populated the rate-limit fields, so a PUT reported
  group ceilings that the following GET omitted; every endpoint now returns them.
- DbUpdateExceptions.IsUniqueViolation replaces the two hand-rolled Postgres
  23505 detectors (ProviderService 43-line walk with repeated magic strings,
  VirtualKeyGroupRepository idempotency-key check).
- ProviderTimeouts.VideoPollingSeconds replaces the duplicated
  CONDUITLLM__TIMEOUTS__VIDEO_POLLING__SECONDS readers in MiniMaxClient and
  OpenRouterClient.
